+++ example/lib/main.dart
@@ -1,7 +1,7 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_plugin_example/application_constants.dart';
 import 'package:flutter_plugin_example/diet_plan.dart';
-import 'package:parse_server_sdk/parse.dart';
+import 'package:parse_server_sdk/parse_server_sdk.dart';
 
 void main() => runApp(new MyApp());
 
@@ -155,9 +155,14 @@ class _MyAppState extends State<MyApp> {
     // Best practice for starting the app. This will check for a valid user
     user = await ParseUser.currentUser();
     await user.logout();
-    user = await ParseUser.currentUser();
 
-    response = await ParseUser.getCurrentUserFromServer();
+    user =
+        ParseUser("TestFlutter", "TestPassword123", "phill.wiggins@gmail.com");
+    response = await user.login();
+    if (response.success) user = response.result;
+
+    response = await ParseUser.getCurrentUserFromServer(
+        token: user.get(keyHeaderSessionToken));
     if (response.success) user = response.result;
 
     response = await user.save();
@@ -178,20 +183,13 @@ class _MyAppState extends State<MyApp> {
   }
 
   function() async {
-    var user =
-        ParseUser("TestFlutter", "TestPassword123", "TestFlutterSDK@gmail.com");
-    await user.signUp();
-    var loginResponse = await user.login();
-    if (loginResponse.success) user = loginResponse.result;
-
-    var customClient = ParseHTTPClient();
-    customClient.additionalHeaders = {
-      keyHeaderSessionToken: ParseCoreData().sessionId
-    };
-    var function = ParseCloudFunction('hello', client: customClient);
-    function.execute();
-
-    user.destroy();
+    var function = ParseCloudFunction('hello');
+    var result = await function.executeObjectFunction<ParseObject>();
+    if (result.success) {
+      if (result.result is ParseObject) {
+        print((result.result as ParseObject).className);
+      }
+    }
   }
 
   functionWithParameters() async {