+++ lib/parse_server_sdk.dart
@@ -22,13 +22,13 @@ part 'src/enums/parse_enum_api_rq.dart';
 
 part 'src/network/parse_http_client.dart';
 
-part 'src/network/parse_livequery.dart';
+part 'src/network/parse_live_query.dart';
 
 part 'src/network/parse_query.dart';
 
 part 'src/objects/parse_base.dart';
 
-part 'src/objects/parse_clonable.dart';
+part 'src/objects/parse_cloneable.dart';
 
 part 'src/objects/parse_config.dart';
 
@@ -58,11 +58,11 @@ part 'src/utils/parse_utils.dart';
 
 class Parse {
   ParseCoreData data;
-  bool _hasBeenInitialised = false;
+  bool _hasBeenInitialized = false;
 
-  /// To initialise Parse Server in your application
+  /// To initialize Parse Server in your application
   ///
-  /// This should be initialised in MyApp() creation
+  /// This should be initialized in MyApp() creation
   ///
   /// ```
   /// Parse().initialize(
@@ -79,6 +79,7 @@ class Parse {
       String clientKey,
       String masterKey,
       String sessionId,
+      bool autoSendSessionId,
       SecurityContext securityContext}) {
     ParseCoreData.init(appId, serverUrl,
         debug: debug,
@@ -87,30 +88,37 @@ class Parse {
         masterKey: masterKey,
         clientKey: clientKey,
         sessionId: sessionId,
+        autoSendSessionId: autoSendSessionId,
         securityContext: securityContext);
 
-    ParseCoreData().initStorage();
-
-    _hasBeenInitialised = true;
+    _hasBeenInitialized = true;
 
     return Parse();
   }
 
-  bool hasParseBeenInitialised() => _hasBeenInitialised;
+  bool hasParseBeenInitialized() => _hasBeenInitialized;
 
-  Future<ParseResponse> healthCheck() async {
+  Future<ParseResponse> healthCheck(
+      {bool debug, ParseHTTPClient client, bool autoSendSessionId}) async {
     ParseResponse parseResponse;
 
+    bool _debug = isDebugEnabled(objectLevelDebug: debug);
+    ParseHTTPClient _client = client ??
+        ParseHTTPClient(
+            autoSendSessionId:
+                autoSendSessionId ?? ParseCoreData().autoSendSessionId,
+            securityContext: ParseCoreData().securityContext);
+
     try {
-      var response = await ParseHTTPClient(ParseCoreData().securityContext)
-          .get("${ParseCoreData().serverUrl}$keyEndPointHealth");
+      var response =
+          await _client.get("${ParseCoreData().serverUrl}$keyEndPointHealth");
       parseResponse =
           ParseResponse.handleResponse(this, response, returnAsResult: true);
     } on Exception catch (e) {
       parseResponse = ParseResponse.handleException(e);
     }
 
-    if (ParseCoreData().debug) {
+    if (_debug) {
       logger(ParseCoreData().appName, keyClassMain,
           ParseApiRQ.healthCheck.toString(), parseResponse);
     }