+++ lib/src/base/parse_constants.dart
@@ -1,12 +1,13 @@
 part of flutter_parse_sdk;
 
 // Library
-const String keySdkVersion = '1.0.10';
+const String keySdkVersion = '1.0.12';
 const String keyLibraryName = 'Flutter Parse SDK';
 
 // End Points
 const String keyEndPointUserName = '/users/me';
 const String keyEndPointLogin = '/login';
+const String keyEndPointLogout = '/logout';
 const String keyEndPointUsers = '/users';
 const String keyEndPointVerificationEmail = '/verificationEmailRequest';
 const String keyEndPointRequestPasswordReset = '/requestPasswordReset';
@@ -21,6 +22,7 @@ const String keyVarUpdatedAt = 'updatedAt';
 const String keyVarUsername = 'username';
 const String keyVarEmail = 'email';
 const String keyVarPassword = 'password';
+const String keyVarSessionToken = 'sessionToken';
 const String keyVarAcl = 'ACL';
 
 // Classes