+++ lib/src/objects/parse_config.dart
@@ -1,12 +1,15 @@
 part of flutter_parse_sdk;
 
 class ParseConfig extends ParseObject {
-  var _client = ParseHTTPClient(ParseCoreData().securityContext);
-
   /// Creates an instance of ParseConfig so that you can grab all configs from the server
-  ParseConfig({bool debug, ParseHTTPClient client}) : super('config') {
-    if (debug != null) setDebug(debug);
-    if (client != null) setClient(client);
+  ParseConfig({bool debug, ParseHTTPClient client, bool autoSendSessionId})
+      : super('config') {
+    _debug = isDebugEnabled(objectLevelDebug: debug);
+    _client = client ??
+        ParseHTTPClient(
+            autoSendSessionId:
+                autoSendSessionId ?? ParseCoreData().autoSendSessionId,
+            securityContext: ParseCoreData().securityContext);
   }
 
   /// Gets all configs from the server